+++ app/helpers/get-valid-session.server.ts
@@ -1,4 +1,4 @@
-import { verifySessionCsrfToken } from "~/components/logic/csrf-token";
+import { verifySessionCsrfToken } from "~/components/logic/csrf-token.server";
 import { getSession } from "~/session.server";
 
 export async function getValidSession(request: Request) {
