+++ app/routes/_layout.charts.$chartName.app-versions.$appVersion.edit.tsx
@@ -16,7 +16,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { getValidSession } from "~/helpers/get-valid-session.server";
 import { useChartAppVersionContext } from "~/routes/_layout.charts.$chartName.app-versions.$appVersion";
