+++ app/routes/_layout.charts.$chartName.edit.tsx
@@ -16,7 +16,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { getValidSession } from "~/helpers/get-valid-session.server";
 import { useChartContext } from "~/routes/_layout.charts.$chartName";
