+++ app/routes/_layout.clusters.$clusterName.($filterNamespace).chart-releases.add.$chartName.tsx
@@ -33,7 +33,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { getValidSession } from "~/helpers/get-valid-session.server";
 import { useSidebar } from "~/hooks/use-sidebar";
 import { commitSession } from "~/session.server";
