+++ app/routes/_layout.clusters.$clusterName.edit.tsx
@@ -27,7 +27,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { getValidSession } from "~/helpers/get-valid-session.server";
 import { useSidebar } from "~/hooks/use-sidebar";
 import { useClusterContext } from "~/routes/_layout.clusters.$clusterName";
