+++ app/routes/_layout.environments.$environmentName.chart-releases.$chartName.change-versions.tsx
@@ -17,7 +17,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { PanelErrorBoundary } from "../errors/components/error-boundary";
 import {
   errorResponseThrower,
