+++ app/routes/_layout.environments.$environmentName.chart-releases.$chartName.edit.tsx
@@ -10,7 +10,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { getValidSession } from "~/helpers/get-valid-session.server";
 import { useEnvironmentChartReleaseContext } from "~/routes/_layout.environments.$environmentName.chart-releases.$chartName";
 import { ChartReleaseEditPanel } from "../features/sherlock/chart-releases/edit/chart-release-edit-panel";
