+++ app/routes/_layout.environments.$environmentName.chart-releases.$chartName.link-pagerduty.tsx
@@ -11,13 +11,13 @@ import {
   ChartReleasesApi,
   PagerdutyIntegrationsApi,
 } from "@sherlock-js-client/sherlock";
-import { getPdAppIdFromEnv } from "~/components/logic/pagerduty-token";
+import { getPdAppIdFromEnv } from "~/components/logic/pagerduty-token.server";
 import { ChartReleaseColors } from "~/features/sherlock/chart-releases/chart-release-colors";
 import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { PanelErrorBoundary } from "../errors/components/error-boundary";
 import {
   errorResponseThrower,
