+++ app/routes/_layout.trigger-incident.$environmentName.general-incident.tsx
@@ -12,7 +12,7 @@ import {
   SherlockConfiguration,
   handleIAP,
 } from "~/features/sherlock/sherlock.server";
-import { formDataToObject } from "~/helpers/form-data-to-object.server";
+import { formDataToObject } from "~/helpers/form-data-to-object";
 import { PanelErrorBoundary } from "../errors/components/error-boundary";
 import { FormErrorDisplay } from "../errors/components/form-error-display";
 import { makeErrorResponseReturner } from "../errors/helpers/error-response-handlers";
