+++ package.json
@@ -1,15 +1,14 @@
 {
+  "type": "module",
   "private": true,
   "sideEffects": false,
   "scripts": {
-    "build": "run-s \"build:*\"",
-    "build:remix": "remix build --sourcemap",
-    "dev": "run-p \"dev:*\"",
-    "dev:remix": "remix dev",
-    "start": "remix-serve ./build/index.js",
+    "dev": "remix vite:dev",
+    "build": "remix vite:build",
+    "start": "remix-serve ./build/server/index.js",
     "typecheck": "npx tsc -b",
     "prepare": "husky",
-    "test": "jest"
+    "test": "vitest"
   },
   "lint-staged": {
     "**/*": "prettier --write --ignore-unknown"
@@ -30,6 +29,7 @@
     "remix-utils": "^7.6.0"
   },
   "devDependencies": {
+    "@mdx-js/rollup": "^3.0.1",
     "@remix-run/dev": "^v2.11.2",
     "@remix-run/eslint-config": "^v2.11.2",
     "@types/jest": "^29.5.12",
@@ -37,13 +37,10 @@
     "@types/react": "^18.3.4",
     "@types/react-dom": "^18.3.0",
     "autoprefixer": "^10.4.20",
-    "cssnano": "^7.0.5",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "husky": "^9.1.5",
-    "jest": "^29.7.0",
     "lint-staged": "^15.2.9",
-    "npm-run-all": "^4.1.5",
     "postcss": "^8.4.41",
     "postcss-cli": "^11.0.0",
     "postcss-fail-on-warn": "^0.2.1",
@@ -52,7 +49,10 @@
     "prettier": "3.3.3",
     "tailwindcss": "^3.4.10",
     "ts-jest": "^29.2.5",
-    "typescript": "^5.5.4"
+    "typescript": "^5.5.4",
+    "vite": "^5.4.2",
+    "vite-tsconfig-paths": "^5.0.1",
+    "vitest": "^2.0.5"
   },
   "engines": {
     "node": ">=20.17.0"
