+++ postcss.config.js
@@ -1,16 +1,14 @@
-module.exports = {
-  plugins: [
+export default {
+  plugins: {
     // Turn @import-glob directives into @import directives
-    require("postcss-import-ext-glob"),
+    "postcss-import-ext-glob": {},
     // Resolve all @import directives
-    require("postcss-import"),
+    "postcss-import": {},
     // Generate Tailwind CSS
-    require("tailwindcss"),
+    tailwindcss: {},
     // Add browser prefixes to CSS to improve compatibility
-    require("autoprefixer"),
-    // Minify everything
-    require("cssnano"),
+    autoprefixer: {},
     // Don't swallow warnings in any of the above
-    require("postcss-fail-on-warn"),
-  ],
+    "postcss-fail-on-warn": {},
+  },
 };
