--- remix.config.js
@@ -1,10 +1,0 @@
-/** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  serverModuleFormat: "cjs",
-  serverDependenciesToBundle: [
-    /^remix-utils.*/,
-    /^@octokit\/.*/,
-    "universal-user-agent",
-    "before-after-hook",
-  ],
-};
