+++ vite.config.ts
@@ -1,0 +1,14 @@
+import mdx from "@mdx-js/rollup";
+import { vitePlugin as remix } from "@remix-run/dev";
+import { installGlobals } from "@remix-run/node";
+import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
+
+installGlobals();
+
+export default defineConfig({
+  server: {
+    port: 3000,
+  },
+  plugins: [mdx(), remix(), tsconfigPaths()],
+});
